// src/app/users-list/model/user.model.ts
export interface User {
    id: number
    name: string
    username: string
    email: string
    website: string
    company: Company
}
export interface Address {
    street: string
    suite: string
    city: string
    zipcode: string
    geo: Geo
}
export interface Company {
    name: string
    catchPhrase: string
    bs: string
}
export interface Geo {
    lat: string
    lng: string
}